mergedFix...

+++ Assets/Scripts/Damagable.cs
@@ -7,13 +7,13 @@ public interface Damagable {
     /// Deal damage to object with given hp.
     /// </summary>
     /// <param name="damageToHp"></param>
-    void damage(int damageToHp);
+    void damage(float damageToHp);
 
     /// <summary>
     /// Heal amount of health to object.
     /// </summary>
     /// <param name="healHp"></param>
-    void heal(int healHp);
+    void heal(float healHp);
 
     /// <summary>
     /// Returns whether true when object has 0 or less health.
